requirements.txt
gradio
torch
torchvision
Pillow
google-generativeai
matplotlib
